backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
anthropic==0.42.0
sentence-transformers==3.3.1
torch==2.5.1
numpy==1.26.4
pandas==2.2.3
datasets==3.2.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
tenacity==9.0.0
tqdm==4.67.1
pytest==8.3.4
pytest-asyncio==0.25.2
aiohttp==3.11.11
